browser/components/backup/tests/chrome/test_backup_settings.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Tests for the BackupSettings component</title>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"></script>
  <script
  src="chrome://browser/content/backup/backup-settings.mjs"
  type="module"
></script>
  <link rel="stylesheet" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
  <script>

    const { BrowserTestUtils } = ChromeUtils.importESModule(
      "resource://testing-common/BrowserTestUtils.sys.mjs"
    );

    /**
     * Tests that adding a backup-settings element to the DOM causes it to
     * fire a BackupUI:InitWidget event.
     */
    add_task(async function test_initWidget() {
      let settings = document.createElement("backup-settings");
      let content = document.getElementById("content");

      let sawInitWidget = BrowserTestUtils.waitForEvent(content, "BackupUI:InitWidget");
      content.appendChild(settings);
      await sawInitWidget;
      ok(true, "Saw BackupUI:InitWidget");

      settings.remove();
    });

    /**
     * Tests that the dialog for turning on scheduled backups can be displayed
     * from settings, or hidden if cancelled.
     */
    add_task(async function test_turnOnScheduledBackupsDialog() {
      let settings = document.getElementById("test-backup-settings");
      settings.backupServiceState = {
        scheduledBackupsEnabled: false,
      }

      await settings.updateComplete;

      let turnOnButton = settings.scheduledBackupsButtonEl;
      let dialog = settings.turnOnScheduledBackupsDialogEl;

      ok(turnOnButton, "Button to turn on scheduled backups should be found");
      ok(!dialog.open, "Dialog should not be open");

      turnOnButton.click();
      await settings.updateComplete;

      ok(dialog?.open, "Dialog should be open");

      let turnOnScheduledBackups = dialog.querySelector("turn-on-scheduled-backups");
      ok(turnOnScheduledBackups, "turn-on-scheduled-backups should be found");

      let cancelButton = turnOnScheduledBackups.shadowRoot.getElementById("backup-turn-on-scheduled-cancel-button");
      ok(cancelButton, "Cancel button should be found");

      cancelButton.click();
      await settings.updateComplete;

      ok(!dialog.open, "Dialog should not be open");
    });
  </script>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
  <backup-settings id="test-backup-settings"></backup-settings>
</div>
<pre id="test"></pre>
</body>
</html>
